compute: move the error scan into its own module
`compute_state.rs` had absorbed a state machine, a scan type, two enums, and
their tests, while the ok phase of the same peek already lived in
`compute_state/peek_result_iterator.rs`. The two phases now sit side by side as
sibling modules, which also turns the `pub(super)` on the cursor type aliases
into a cross-module contract rather than a private alias pried open by the
parent.

Pure move. The only edits are the visibility markers the module boundary forces
and two intra-doc links that now have to leave their defining module.
